+++ src/common/CMakeLists.txt
@@ -175,7 +175,11 @@ set( SOURCE_FILES ${COMMON_BASE_HEADERS} ${COMMON_BASE_SOURCES} )
 source_group( common FILES ${COMMON_BASE_HEADERS} ${COMMON_BASE_SOURCES} )
 add_library( common_base ${SOURCE_FILES} )
 #message( STATUS "common_base LIBRARIES=${LIBRARIES}, DEFINITIONS=${DEFINITIONS}")
-target_link_libraries( common_base ${LIBRARIES} gtest_main gmock_main )
+if ( TESTING )
+	target_link_libraries( common_base ${LIBRARIES} gtest_main gmock_main )
+else()
+	target_link_libraries( common_base ${LIBRARIES} )
+endif()
 set_target_properties( common_base PROPERTIES COMPILE_FLAGS "${DEFINITIONS}" )
 include_directories( ${INCLUDE_DIRS} )
