+++ src/common/core.cpp
@@ -16,6 +16,9 @@
 #include "winapi.hpp" // Console close event handling
 #include <direct.h> // _chdir
 #endif
+#ifdef TESTING
+#include "gtest/gtest.h"
+#endif
 
 #include "cbasetypes.hpp"
 #include "malloc.hpp"
@@ -337,7 +340,10 @@ int main (int argc, char **argv)
 	}
 
 	malloc_init();// needed for Show* in display_title() [FlavioJS]
-
+#ifdef TESTING
+	::testing::InitGoogleTest(&argc, argv);
+	return RUN_ALL_TESTS();
+#endif
 #ifdef MINICORE // minimalist Core
 	display_title();
 	usercheck();
