+++ src/map/CMakeLists.txt
@@ -76,7 +76,11 @@ source_group( map FILES ${MAP_HEADERS} ${MAP_SOURCES} )
 include_directories( ${INCLUDE_DIRS} )
 
 #message( STATUS "map-server SOURCE_FILES=${SOURCE_FILES}")
-add_executable( map-server ${SOURCE_FILES} ${TEST_SOURCES} )
+if ( TESTING )
+	add_executable( map-server ${SOURCE_FILES} ${TEST_SOURCES} )
+else()
+	add_executable( map-server ${SOURCE_FILES} )
+endif()
 #message( STATUS "map-server LIBRARIES=${LIBRARIES}, DEPENDENCIES=${DEPENDENCIES} DEFINITIONS=${DEFINITIONS}")
 add_dependencies( map-server ${DEPENDENCIES} )
 target_link_libraries( map-server ${LIBRARIES} ${DEPENDENCIES} )
