+++ src/map/script.hpp
@@ -3,11 +3,12 @@
 
 #ifndef SCRIPT_HPP
 #define SCRIPT_HPP
-
+#include <memory>
 #include "../common/cbasetypes.hpp"
 #include "../common/db.hpp"
 #include "../common/mmo.hpp"
 #include "../common/timer.hpp"
+#include "map_interface.hpp"
 
 #define NUM_WHISPER_VAR 10
 
@@ -1988,6 +1989,6 @@ bool script_check_RegistryVariableLength(int pType, const char *val, size_t* vle
  **/
 typedef int (*script_func)(struct script_state *st);
 script_func get_func_ptr(int id);
-//void script_set_map_obj(Map_Obj map_obj);
+void script_set_map(std::shared_ptr<Map_Interface> map_obj_);
 
 #endif /* SCRIPT_HPP */
